Disk paths are resolved perfectly.
Corrected an issue where dots followed immediately by a char would pass. Now error in these cases.

UNC paths are resolved according to the standard.
Improved DOS compatibility with UNC resolution, so it is very similar to DOS and way more deterministic.

Possibly will remove the special case of the "suppression" of the machine name trailing path separator, but in for now for better DOS compatibility.