Added DOS 4 functionality to get the true version and revision numbers from 41/3306h, get 1 based boot drive from 41/3305h and set reporting version number in 4F/122Fh. All these functions will remain undocumented for now (though we can use these services when it is vital to do so only. Otherwise, please use normal services or avoid them altogether, to avoid drawing attention to them).
Made the DOS version number a macro to encode the true version in the instructions to prevent easily changing the reporting number. Adding it as a macro means that we can also pass a different version and rev number from a linker too!
The variable which can be changed is initialised to this macro value.