Remodelled the IO buffer handling for file write to defer writes to mass storage until the absolute last moment possible. Now we make use of the referenced bit to indicate which buffers we have made use of during a write.
All buffers have their ref bit set when obtained. We clean this on all buffers when entering Read/Write disk bytes. Thus we can rely on ref and dirty bits only to know which dirty buffers to throw away if a hard error occurs! SFT not updated on hard error, as all written data is thrown away. It is possible that at the end of a write, nothing has been written to disk, and only lives in buffers. As usual, dir information only updated on handle close.
This change does not hold for mkdir/rmdir, as they immediately are written back to the disk.