Fixed a couple of esoteric bugs in DOS that affected find next! findNextMain now more robustly sets up the dos internal variables, building a copy of the CDS instead in the SDA.
Also adjusted byte 0 of the FF block to be the 1 based drive number, not 0 based as find first handle was leaving it in.

Caught a bug in walkDPB that started walking from the sft pointer (doh!)

Fixed comments here and there.

We now set the archive bit correctly on written files (before, only create would set it).

Improved get/set filetime to now work properly with share and critical sections around SFT access!