Fixed a join bug in the find file that would result in not getting the right DPB ptr when not doing successive calls on consecutive directories on a join drive.
Fixed a bug with net qualify, that would always exit when "hitting" the disk for a net file upon name qualification. This would cause a bug with reading volume labels and finding files using FCB calls to drive-leter mounted network drives.

I suspect the join issue is also a result of the fact that we now use an FCB calls in the DIR command, and that handle based access would continue to not cause any issues.