Set find first to now only return devices when doing file or dir searches. Will not return devices if doing exclusive dir searches. This may allow for the removal of the safety guards in the dir functions.
Now the dev\chardev intervention works properly and resolves names exactly like DOS!

Also find next fails on char devices, but now we have two reasons for it to fail, one being that the fnd attribute was 40h and the second is that the directory offset is -1 which is invalid. Both are escape conditions to protect in the event of a call with a damaged ffblock.