Adjusted a buffer funciton to use a macro for quick returns.
Adjusted comments in the code for 21/53 to make clear what we are computing.

Ensure in getDiskData we return the number of clusters on disk and not the maximum cluster address.

Fixed free cluster reporting bugs in findFreeClusterData when there are no free clusters. If it was never asked to report the freeclusters after the dpb was built, it may return -1 when there are 0 free clusters.

Fixed FAT 12 reading and writing FAT entries cross sector boundary. Corrected an incorrect assumption that even clusters cannot go across a sector boundary.

Writes which extend/truncate a file are now being exhaustively tested (alongside normal writes) to ensure that the correct number of bytes are reported in the directory entry. Disk full conditions are now correctly handled and reported if necessary. Int 24h on disk full is now ready to be implemented.

writeDiskFile is now seemingly capable of handling properly the case of 0 byte writes, as well as partial writes, starting from any offset (including out of the file) and writing with bytes as well as truncating files. Need some more testing but at least extension and normal writes have been tested and confirmed working.

May want to add: CTRL+C checking during extension. Could lead to lost clusters.

Breakpoints all commented out!