For "compatibility" with DOS 3.3, we added the null marker pointer to the share hook table and modified the sharehook setup. No programs (yet) make use of the sharehooks so no modifications necessary (as all programs which use the SDA, go via the doscall). Removed the "simpler" call that I had added to be more similar to FreeDOS. Now we are back to being more like MS.
Furthermore, major changes to the routine that checks Exclusive Ownership of a File by a process for sharing purposes to be simpler and work almost exactly the same as in DOS. Due to subtle differences in the way we do things internally, our approach differs from MS but as far as the Sharing utility is concerned, all the same info is passed. All working ok!