Ok edits made, now need to debug.
Added a pseudo function call that allows for closing FCBs. The only FCBs which can be opened "autoclose" for now. This is so that software which writes volume labels using the FCB endpoint can be "future proofed" as we never admit that the open/create FCB call on labels autocloses.