Now we also compute the number of cylinders on the media described by the BPB and store it in the drvblk. This is not dependable on removable media and should only be used for fixed media where the BPB doesnt change and the dfltBPB and actualBPB are equal.
Wrapped the setting of the registers before an IO call into a function so that blkIO can be called also by IOCTL calls. with registers carrying the values into blkIO.

Also fixed a bug that would occur if more than one sector was to be read.

We now preserve all the registers we need to preserve by always calling Int 33h wrapped by pushes and pops of the registers we need preserved across the call. Realistically not necessary as BIOS preserves regs but I cant seem to guarantee that all necessary regs are preserved across all EHCI calls. Really need a new, better written BIOS.