Fixed a minor bug where logical partitions would be marked as primary in the drvBlk (this is never checked).
Added a function that now takes the information about partition size and start sector from the MBR rather than the BPB. This is used to setup these values in drvBlk for partitions that are unformatted. For formatted partitions, these values are still taken from the BPB so as to not violate if a "smaller" BPB has been layed onto a "larger" partition.