Moved equates and structs local to the msd driver into the include file but I dont expose these as global symbols.
Added special handling for when no valid partitions are found. We don't turn off the system but proceed hoping a default drive A works.

Also now all drvtbl bpb entries are initialised to a 1.44Mb default layout so that when an unformatted partition is encountered, the values except for .dHiddSec and .w/dTotSec16/32 refer to this default layout, and we set the unformatted bit in the flags. This is so that a valid non-buggy DPB is built for the drive, but the driver will refuse to operate on this drive until a format utility has formatted this drive.