Added support for pre DOS 2 FAT12 filesystems (i.e. those without a BPB).
Fixed medcheck/buildBPB to now actually change BPB if a device change occured, unlike before where it was returning prematurely.

Added more flag setting to msdinit.

Improved a buffer function.

Left to check if the IOCTL functions will suffice for FDISK/FORMAT purposes. This will be the subject of the next commit :)