Moved the setting of biosVers and biosPtr to just before kernel driver init so that if OEMINIT cannot pass arguments and/or pointers through known internal variables to the Kernel drivers, they can use the memory location in DOS to do so. This pointer also allows later installable drivers to directly access BIOS services if the driver recognises the BIOS version without which the pointer is formally meaningless.
For us, this pointer will serve as a UEFI HAL replacement entrypoint for the BIOS services/lookup table of HAL services, thus freeing the interrupts and also shielding the HAL from most users (as the biosPtr var will never be documented for use).

On SCPBIOS systems, the BIOS vers is 0 and the biosPtr is 0.