Fixed FAT 32 booting error.
BuildDPB on FAT32 needs a sector in the buffer chain to read the FSInfo sector into. This was being allocated but never linked into the buffer chain, nor was it ever freed. Now restructured the code so that a buffer is ONLY allocated for kernel disk driver loading, linked into the buffer chain and free'd immediately afterw DPBs have been built for the kernel disk driver. Config.sys processing of disk drives is still done as before, since at that point there is a single disk buffer in the system.

Split the error handling code into three smaller files.