Added get/set disk serial number, both at the DOS level and at the disk driver level. Tested and seems to be working a charm.
Corrected a minor oversight in rename with wildcards. Any *.* rename in the same directory would infinite loop as we would always create a new directory entry in advance of the current search pointer, and thus every new file would get deleted and a new file would be created.
Now we check, if we are in the same dir, we recycle the directory entry of the file being renamed. If we are changing dirs, this never causes any problems and so we do things properly. WC copy now confirmed ok. WC delete too.

Made some improvemens to the disk driver (improved error handling and made sure that block IO always correctly reports errors).