FIXED BUG WHERE COPIES ON SHARED A:-B: WERE FAILING.
CURRENTSFT WAS BEING RESET ON ALL ENTRIES TO INT21! NOW NEVER RESET IT (each handle function sets the SFT as needed, including delete, rename and chmod).

Commented out "checkDoInt24OnHandle" and the clearing of currentSFT in init since firstly, that bit cannot yet be set on handles and secondly, its presence is what caused aforementioned bug in the first place... will think of a more elegent way to do that :)

Just cleaned up a lot of comments and code.

Started reworking the debug features, might add them globally and permanently across the kernel to aid with error handling when working on physical machines.