Improved rename a bit further. Now you cannot rename read-only files.
Also finally restructured it so it is easier to follow the WC loop.

Like with Delete, consider moving the flush operation to the end, so that in a WC loop, we dont flush after each file.

The actual renaming subroutines have recieved some improvements and a bug fix (renames in the same dir could rename a file to a file that already exists, this has been patched).