Changed sharehook names and added some more symbols and better comments for fields in the sft structure.
buildSFTEntry now also sets dAbsClust to dStartClust and dRelClust to 0 when creating/opening disk files. We still don't rely on these entries yet in the dos kernel.

Consider adding dependency on them when they are not start of file markers or are and dCurtOff is 0.

Implemented updateDirShare to ensure that all SFTs belonging to an MFT get updated when a dir info changes for any SFT in the chain. These also modify the cluster data in the SFTs in anticipation of the future upgrade of the DOS kernel.