Fixed a bug in 21h/31h where a the register with the psp segment wasnt being correctly preserved across a call
Added some comments to clean up the hdlfunctions.

Now when closing we make a copy of the current directory as we will write it to disk to the curdircopy in the sda so truncate has a correct copy of the dir to use in a multitasking environment.

Added a variable in the data area to keep track of the the original DOS PSP (needed for sharing, to locate the handles belonging to DOS's startup PSP).

Implemented closeRenDel and closeNetworkFiles. The latter cant really be tested but can be visually verified to do the right thing.
The former has been thoroughly tested to be working. A crash was present but its fix seems to be related to the to Del infinitely looping (Needed to preserve rbx across a 2Fh/1216h call)